SWDEV-505971 - change setArgument arg from uint32_t to uint64_t
We are passing this arg as an address, and memcpy complains about
overreading (8 bytes instead of 4).



[ROCm/clr commit: a6f2a2c2afa388e69bc7389622dfec480a2f35bb]

+++ rocclr/device/blit.hpp
@@ -236,7 +236,7 @@ class BlitManager : public amd::HeapObject {
   //! Stream batch memory operation
   virtual bool batchMemOps(const void* paramArray,
                            size_t paramSize,
-                           uint32_t count
+                           uint64_t count
   ) const = 0;
 
   //! Enables synchronization on blit operations
